Added the GenSec Security Website
Most of this work was done today, on sunday. Including the Contact/404 website, the entire News site, and my javascript shenanigans. It's all a bit messy but getting to use JS for the first time was fun.
Am also so happy the entire buttons on the menu are now clickable, for the longest time you could only click the text/link to switch between the pages.